% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zeros_count.R
\name{add_zeros_count}
\alias{add_zeros_count}
\title{Add in zeros for deployments not present in the count data}
\usage{
add_zeros_count(count_data, metadata)
}
\arguments{
\item{count_data}{A data frame of species counts.}

\item{metadata}{A data frame containing deployment-level metadata.}
}
\value{
A dataframe with a row for every species in the count data, in every sample in the metadata
}
\description{
Add in zeros for deployments not present in the count data
}
