% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_catami.R
\docType{data}
\name{catami}
\alias{catami}
\title{CATAMI classification reference table}
\format{
A data frame describing CATAMI class names/codes and any hierarchical
relationships or descriptors included in the CheckEM distribution.
}
\source{
CATAMI classification framework as compiled for use in CheckEM.
}
\usage{
catami
}
\description{
A reference table containing CATAMI (Collaborative and Automated Tools for
Analysis of Marine Imagery) classes and/or mappings used by CheckEM.
}
\details{
Used to support habitat / benthic classification validation and standardisation.
}
\keyword{datasets}
