% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{A function to standardise names in a data frame}
\usage{
clean_names(dat)
}
\arguments{
\item{dat}{}
}
\value{
A data frame with all lower case column names. This function also replaces any special characters with an "_".
}
\description{
A function to standardise names in a data frame
}
