% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_count.R
\name{complete_count}
\alias{complete_count}
\title{Add in zeros for deployments not present in the count data}
\usage{
complete_count(count_data, metadata)
}
\arguments{
\item{count_data}{}

\item{metadata}{}
}
\value{
A dataframe with a row for every species in the count data, in every sample in the metadata
}
\description{
Add in zeros for deployments not present in the count data
}
