% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_length.R
\name{complete_length}
\alias{complete_length}
\title{Add in zeros for deployments not present in the length data}
\usage{
complete_length(length_data, metadata)
}
\arguments{
\item{length_data}{}

\item{metadata}{}
}
\value{
A dataframe with a row for every species in the length data, in every sample in the metadata
}
\description{
Add in zeros for deployments not present in the length data
}
