% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_min_max.R
\name{create_min_max}
\alias{create_min_max}
\title{A function to create minimum and maximum size limits to check annotations against}
\usage{
create_min_max(life_history, minimum, maximum)
}
\arguments{
\item{life_history}{}
}
\value{
A life history dataframe, either australia_life_history or global_life_history
}
\description{
A function to create minimum and maximum size limits to check annotations against
}
