% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_life_history.R
\name{expand_life_history}
\alias{expand_life_history}
\title{A function to expand a life-history dataframe so there is one row per species per region}
\usage{
expand_life_history(dat)
}
\arguments{
\item{dat}{}
}
\value{
A life history dataframe, either australia_life_history or global_life_history
}
\description{
A function to expand a life-history dataframe so there is one row per species per region
}
