% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tl.R
\name{find_tl}
\alias{find_tl}
\title{Retrieve Trophic Level and Standard Error from FishBase}
\usage{
find_tl(sp, mirror = "us")
}
\arguments{
\item{sp}{A character string representing the species name. The species name should be provided
as it appears on FishBase, but spaces will be replaced with hyphens.}

\item{mirror}{A character string specifying the FishBase mirror site to use. Defaults to "us"
(FishBase USA). Other mirrors can be specified if needed.}
}
\value{
A data frame containing the trophic level information for the species. The data frame
includes the following columns:
\itemize{
\item \strong{species}: The species name.
\item \strong{trophic_level}: The trophic level of the species.
\item \strong{se}: The standard error of the trophic level.
\item \strong{metric}: The source used (if provided) e.g. "based on food items".
}
If there is no internet connection, or if the species page cannot be retrieved or parsed, the function
returns \code{NULL}.
}
\description{
This function retrieves the trophic level and its standard error for a given species from FishBase.
It fetches data from the FishBase website, extracts the trophic level information, and formats it
into a data frame.
}
\examples{
# Retrieve trophic level and standard error for a species
tl_info <- find_tl("Lutjanus campechanus")
print(tl_info)

}
