% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_api_all_data.R
\name{ga_api_all_data}
\alias{ga_api_all_data}
\title{Retrieve all Data from a synthesis using the GlobalArchive API}
\usage{
ga_api_all_data(token, synthesis_id, dir, include_zeros = FALSE)
}
\arguments{
\item{token}{A character string representing your GlobalArchive token for API authentication.}

\item{synthesis_id}{A character string or numeric value representing the GlobalArchive synthesis ID
for which the data should be retrieved.}

\item{dir}{A character string specifying the directory where the RDS files will be saved.}

\item{include_zeros}{Either TRUE or FALSE, if you would like the data to include zeros where a species is not observed, warning: this can create large datasets that use a lot og RAM}
}
\value{
A list containing \code{metadata}, \code{count}, and \code{length} data frames.
}
\description{
This function retrieves metadata, count, and length data from a CheckEM synthesis, processes the data,
and saves the results as RDS files in the specified directory with the given name. It also returns the
processed data in the environment and allows users to add in zeros where a species is not present
If the directory doesn't exist, it will be created, and the user will be informed.
}
