% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_api_length.R
\name{ga_api_length}
\alias{ga_api_length}
\title{Function to API call length data from a synthesis in GlobalArchive}
\usage{
ga_api_length(username, password, synthesis_id)
}
\description{
Function to API call length data from a synthesis in GlobalArchive
}
