% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_api_set_token.R
\name{ga_api_set_token}
\alias{ga_api_set_token}
\title{Set GlobalArchive API Token}
\usage{
ga_api_set_token()
}
\description{
This function prompts the user to enter their GlobalArchive API token, saves it in an RDS file
within a "secrets" directory, and optionally adds the "secrets" directory to
the \code{.gitignore} file if the function is run inside a Git repository.
}
\details{
The function performs the following steps:
\itemize{
\item Prompts the user to input their GlobalArchive API token via the console.
\item Creates a "secrets" directory in the current working directory if it doesn't already exist.
\item Saves the token as an RDS file named \code{api_token.rds} inside the "secrets" directory.
\item Checks if the current directory is part of a Git repository (i.e., it contains a \code{.git} folder).
\item If in a Git repository, the function adds the "secrets/" directory to the \code{.gitignore} file if it's not already listed.
}
}
\note{
This function is intended to be used in projects involving GlobalArchive, where sensitive information
like API tokens should not be committed to version control. Ensure that the working directory is set
correctly before running this function.
}
\examples{
\dontrun{
  # Run the function to set a GlobalArchive API token
  set_token()
}

}
