% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_api_species_list.R
\name{ga_api_species_list}
\alias{ga_api_species_list}
\title{Retrieve Species List from the Global Archive API}
\usage{
ga_api_species_list(token)
}
\arguments{
\item{token}{A character string representing your GlobalArchive token for API authentication.}
}
\value{
A data frame containing the species list from GlobalArchive.
The data frame includes the following columns:
\itemize{
\item \code{subject}: The URL or identifier of the species in GlobalArchive.
\item \code{australian_common_name}: Common name of the species.
\item \code{family}: The family of the species.
\item \code{genus}: The genus of the species.
\item \code{species}: The species name.
\item \code{caab}: The CAAB (Codes for Australian Aquatic Biota) code.
\item \code{other_columns}: Any other relevant columns included in the original dataset.
}
}
\description{
This function retrieves the species list from GlobalArchive by making an API call. The data
is returned as a data frame with relevant species information, excluding some columns that
are not needed for further analysis.
}
\examples{
\dontrun{
# Fetch the species list from GlobalArchive
species_list <- ga_api_species_list("your_username", "your_password")

# Display the first few rows of the species list
head(species_list)
}
}
