% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rds_from_github.R
\name{load_rds_from_github}
\alias{load_rds_from_github}
\title{Load an RDS File from a Private GitHub Repository}
\usage{
load_rds_from_github(url, pat = NULL)
}
\arguments{
\item{url}{Raw URL to the file (raw.githubusercontent.com/...)}

\item{pat}{GitHub Personal Access Token. If NULL, tries env vars then gitcreds.}
}
\value{
R object, or NULL if not available (non-interactive w/out PAT) / access denied.
}
\description{
Load an RDS File from a Private GitHub Repository
}
\examples{
\dontrun{
rds_url <- "https://raw.githubusercontent.com/GlobalArchiveManual/australia-synthesis-2024/main/data/tidy/australian-synthesis_covariates.RDS"
data <- load_rds_from_github(rds_url)
}
}
