% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maturity.R
\docType{data}
\name{maturity}
\alias{maturity}
\title{Maturity reference table}
\format{
A data frame with one row per taxon (or taxon-region combination) and
columns describing maturity metrics or categories.
}
\source{
Compiled by the CheckEM team from maturity reference sources.
}
\usage{
maturity
}
\description{
A reference dataset of maturity-related information (e.g. length-at-maturity or
maturity status categories) used by CheckEM.
}
\details{
Used for QA/QC checks and/or derived summaries where maturity information is relevant.
}
\keyword{datasets}
