% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_hdi.R
\name{plot_species_hdi}
\alias{plot_species_hdi}
\title{Plot length frequency histogram with High Density Interval}
\usage{
plot_species_hdi(species_data, max_lengths, dir)
}
\arguments{
\item{species_data}{A data frame with length information for one species}

\item{max_lengths}{A data frame with lmaximum length information}

\item{dir}{Directory where plots will be saved}
}
\value{
Saves plots into directory
}
\description{
This function takes a data frame and standardises its column names by converting them to lower case,
replacing special characters with underscores, and ensuring that names are syntactically valid R names.
}
