% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_counts.R
\name{read_counts}
\alias{read_counts}
\title{Read Count Data from Generic Count Data (_Count.csv)}
\usage{
read_counts(dir, method = "BRUVs", recursive = FALSE)
}
\arguments{
\item{dir}{A character string specifying the directory containing the count files.}

\item{method}{A character string specifying the method used: either \code{"BRUVs"} or \code{"DOVs"}.
Defaults to \code{"BRUVs"}.}

\item{recursive}{A logical value indicating whether to read files recursively from the directory.
Defaults to \code{FALSE}.}
}
\value{
A data frame containing the cleaned and formatted count data from the specified files.
}
\description{
This function reads and processes point count data from Generic files
in either CSV or TXT format. It can handle both BRUVs (Baited Remote Underwater Video Systems)
and DOVs (Diver Operated Video) methods, ensuring that the data is correctly cleaned and
formatted for further analysis.
}
\examples{
\dontrun{
# Read all count files from the directory for BRUVs method
counts_data <- read_counts(dir = "path/to/directory", method = "BRUVs", recursive = TRUE)

# Read all count files from the directory for DOVs method
counts_data <- read_counts(dir = "path/to/directory", method = "DOVs", recursive = TRUE)
}
}
