% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_em_length.R
\name{read_em_length}
\alias{read_em_length}
\title{Read EventMeasure length and 3D point data (_3DPoints.txt & _Lengths.txt)}
\usage{
read_em_length(dir, method = "BRUVs", recursive = FALSE)
}
\arguments{
\item{dir}{The directory where the EventMeasure length data files are located.}

\item{method}{A character string indicating the method used ("BRUVs" or "DOVs"). Default is "BRUVs".}

\item{recursive}{Logical, whether to search for files recursively in subdirectories. Default is \code{FALSE}.}
}
\value{
A data frame containing all length data from the specified files, with standardised column names and formats.
}
\description{
This function reads length and 3D point information from EventMeasure export files (\verb{_3DPoints.txt} and \verb{_Lengths.txt}) from a specified directory
and processes them into a single dataframe. The dataframe includes details such as length, species, sample, and additional metadata.
}
\examples{
\dontrun{
# To read EventMeasure length data from a directory using the BRUVs method
length_data <- read_em_length("path/to/directory")

# To read length data from a directory recursively using the DOVs method
length_data <- read_em_length("path/to/directory", method = "DOVs", recursive = TRUE)
}
}
