% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_length.R
\name{read_files_length}
\alias{read_files_length}
\title{A function to read EventMeasure format length data (_3DPoints.txt & _Lengths.txt)}
\usage{
read_files_length(dir)
}
\arguments{
\item{dir}{}
}
\description{
A function to read EventMeasure format length data (_3DPoints.txt & _Lengths.txt)
}
