% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gen_length.R
\name{read_gen_length}
\alias{read_gen_length}
\title{A function to read "Generic" format length data _Length.csv}
\usage{
read_gen_length(dir, method = "BRUVs")
}
\arguments{
\item{dir}{}
}
\description{
A function to read "Generic" format length data _Length.csv
}
