% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read Metadata}
\usage{
read_metadata(dir, method = "BRUVs", recursive = FALSE)
}
\arguments{
\item{dir}{The directory where the .csv files are saved.}

\item{method}{A character string indicating the method used ("BRUVs" or "DOVs"). Default is "BRUVs".}

\item{recursive}{Logical, whether to search for files recursively in subdirectories. Default is \code{FALSE}.}
}
\value{
A data frame containing all metadata from the specified files, with standardised columns
for campaignID, sample, and various other metadata details.
}
\description{
This function reads Metadata in a GlobalArchive format (as .csv files) from a specified directory
and processes them into a single dataframe. The dataframe includes campaignID, sample, depth_m, time, coordinates and other important information collected when deploying and annotating stere-video imagery.
}
\examples{
\dontrun{
# To read metadata from a directory using the BRUVs method
metadata <- read_metadata("path/to/directory")

# To read metadata from a directory recursively using the DOVs method
metadata <- read_metadata("path/to/directory", method = "DOVs", recursive = TRUE)
}
}
