% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_points.R
\name{read_points}
\alias{read_points}
\title{A function to read EventMeasure format point data (_Points.txt)}
\usage{
read_points(dir)
}
\arguments{
\item{dir}{}
}
\description{
A function to read EventMeasure format point data (_Points.txt)
}
