% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheckEM.R
\name{runCheckEM}
\alias{runCheckEM}
\title{Run the CheckEM Shiny App Locally}
\usage{
runCheckEM()
}
\value{
Launches the CheckEM Shiny app in the web browser.
}
\description{
This function launches a local version of the CheckEM Shiny app in your default web browser.
The app allows users to perform quality control checks on metadata and annotations for fish
and benthic stereo-imagery datasets.
}
\examples{
\dontrun{
# To run the CheckEM Shiny app locally
CheckEM::runCheckEM()
}
}
