% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_tibble.R
\name{standardize_tibble}
\alias{standardize_tibble}
\title{Function to standardize column names and types}
\usage{
standardize_tibble(tibble)
}
\arguments{
\item{tibble}{}
}
\description{
Function to standardize column names and types
}
