% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{A function to standardise names in a dataframe}
\usage{
clean_names(dat)
}
\arguments{
\item{dat}{}
}
\value{
A dataframe with all lower case names, with _ as separators
}
\description{
A function to standardise names in a dataframe
}
