% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_api_count.R
\name{ga_api_count}
\alias{ga_api_count}
\title{Function to API call count data from a synthesis in GlobalArchive}
\usage{
ga_api_count(username, password, synthesis_id, include_life_history = TRUE)
}
\description{
Function to API call count data from a synthesis in GlobalArchive
}
