% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_api_metadata.R
\name{ga_api_metadata}
\alias{ga_api_metadata}
\title{Function to API call metadata from a synthesis in GlobalArchive}
\usage{
ga_api_metadata(username, password, synthesis_id)
}
\description{
Function to API call metadata from a synthesis in GlobalArchive
}
