% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_api_species_list.R
\name{ga_api_species_list}
\alias{ga_api_species_list}
\title{Function to API call to the GlobalArchive species list}
\usage{
ga_api_species_list(username, password)
}
\description{
Function to API call to the GlobalArchive species list
}
