% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_special_characters.R
\name{has_special_characters}
\alias{has_special_characters}
\title{Function to check if a string contains any special characters}
\usage{
has_special_characters(string)
}
\arguments{
\item{string}{}
}
\description{
Function to check if a string contains any special characters
}
