% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transformations.R
\name{plot_transformations}
\alias{plot_transformations}
\title{A function to test basic transformations on predictor variables for GAM modelling}
\usage{
plot_transformations(pred.vars, dat)
}
\arguments{
\item{dat}{}
}
\description{
A function to test basic transformations on predictor variables for GAM modelling
}
