% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_counts.R
\name{read_counts}
\alias{read_counts}
\title{A function to read EventMeasure format point data (_Points.txt)}
\usage{
read_counts(dir, method = "BRUVs", recursive = FALSE)
}
\arguments{
\item{dir}{}
}
\description{
A function to read EventMeasure format point data (_Points.txt)
}
