% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_em_length.R
\name{read_em_length}
\alias{read_em_length}
\title{A function to read EventMeasure format length data (_3DPoints.txt & _Lengths.txt)}
\usage{
read_em_length(dir, method = "BRUVs", recursive = FALSE)
}
\arguments{
\item{dir}{}
}
\description{
A function to read EventMeasure format length data (_3DPoints.txt & _Lengths.txt)
}
