% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{A function to read GlobalArchive format metadata}
\usage{
read_metadata(dir)
}
\arguments{
\item{dir}{The directory where the .txt files are saved}
}
\value{
A dataframe with all a column for campaignID and sample
}
\description{
A function to read GlobalArchive format metadata
}
