% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_TM.R
\name{read_TM}
\alias{read_TM}
\title{A function to read .txt exports from SeaGIS TransectMeasure}
\usage{
read_TM(dir, sample)
}
\arguments{
\item{dir}{The directory where the .txt files are saved}

\item{sample}{Either character value to define if sample is recorded as "opcode" or "period"}
}
\value{
A dataframe with all a column for campaignID and sample
}
\description{
A function to read .txt exports from SeaGIS TransectMeasure
}
