% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheckEM.R
\name{runCheckEM}
\alias{runCheckEM}
\title{Run the CheckEM Shiny App Locally}
\usage{
runCheckEM(
  install_missing = TRUE,
  ask = interactive(),
  repos = getOption("repos")
)
}
\arguments{
\item{install_missing}{Logical. If TRUE, install missing packages automatically.}

\item{ask}{Logical. If TRUE, ask before installing packages.}

\item{repos}{Character. CRAN repo to use.}
}
\value{
Launches the CheckEM Shiny app.
}
\description{
Launches the CheckEM Shiny app in the default web browser. Optionally installs
required packages if they are missing.
}
\examples{
\dontrun{
CheckEM::runCheckEM()
}
}
