% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheckEM.R
\name{runCheckEM}
\alias{runCheckEM}
\title{Run CheckEM shiny app locally}
\usage{
runCheckEM()
}
\value{
Will launch a local version of CheckEM in the browser
}
\description{
Run CheckEM shiny app locally
}
\examples{
## Not run:
\dontrun{CheckEM::runCheckEM()}

}
